refactor: supabase auth service
